Simplify the way we handle formalist serialisation, now on submit only.
We really only care about having the hidden serialised input being
present just before we submit the form. The rest of the time it’s just
thrashing the layout, and so it’s totally pointless. This change removes
the JIT updates, and defers the render until the form is actually
submitted (which we were doing anyway).